// src/faker/data-type.enum.ts
export enum DataType {
  FAKER_PERSON_NAME = 'FAKER_PERSON_NAME',
  FAKER_FIRST_NAME = 'FAKER_FIRST_NAME',
  FAKER_LAST_NAME = 'FAKER_LAST_NAME',

  FAKER_EMAIL = 'FAKER_EMAIL',
  FAKER_USERNAME = 'FAKER_USERNAME',
  FAKER_PASSWORD = 'FAKER_PASSWORD',

  FAKER_AVATAR = 'FAKER_AVATAR',
  FAKER_IMAGE_URL = 'FAKER_IMAGE_URL',

  FAKER_PHONE = 'FAKER_PHONE',

  FAKER_UUID = 'FAKER_UUID',

  FAKER_CITY = 'FAKER_CITY',
  FAKER_COUNTRY = 'FAKER_COUNTRY',
  FAKER_STREET = 'FAKER_STREET',
  FAKER_ZIPCODE = 'FAKER_ZIPCODE',

  FAKER_COMPANY = 'FAKER_COMPANY',
  FAKER_JOB_TITLE = 'FAKER_JOB_TITLE',

  FAKER_PRODUCT = 'FAKER_PRODUCT',
  FAKER_PRICE = 'FAKER_PRICE',

  FAKER_SENTENCE = 'FAKER_SENTENCE',
  FAKER_PARAGRAPH = 'FAKER_PARAGRAPH',

  FAKER_NUMBER = 'FAKER_NUMBER',
  FAKER_BOOLEAN = 'FAKER_BOOLEAN',

  FAKER_DATE_PAST = 'FAKER_DATE_PAST',
  FAKER_DATE_FUTURE = 'FAKER_DATE_FUTURE',
}